requirements.txt
# Core RAG & Vector Database
langchain==0.3.15
langchain-community==0.3.15
chromadb==0.5.23
sentence-transformers==3.3.1

# LLM API (DeepSeek via DeepInfra - OpenAI compatible)
openai==1.58.1

# Web Framework & UI
streamlit==1.41.1

# Data Processing (numpy<2 is required by langchain)
pandas==2.2.3
numpy==1.26.4
pypdf==5.1.0
python-dotenv==1.0.1

# Utilities
tiktoken==0.8.0
requests==2.32.3